Modularize the code base.
This is still the initial step where the scanner and the parser are
separated into two big chunks (with utility modules on the side as
well). It is expected that the coupling between the two will be better
defined in the subsequent refactoring.

Closes #1176
