+++ .gitignore
@@ -1,3 +1,4 @@
 coverage
 node_modules
 test/dist
+src/*.js
