+++ package.json
@@ -57,6 +57,7 @@
     "temp": "~0.8.3",
     "tslint": "~2.5.1",
     "typescript": "~1.6.2",
+    "typescript-formatter": "~1.1.0",
     "unicode-7.0.0": "~0.1.5",
     "webpack": "~1.12.1"
   },
@@ -71,7 +72,7 @@
   "scripts": {
     "check-version": "node test/check-version.js",
     "tslint": "tslint src/*.ts",
-    "code-style": "jscs -p crockford test/*.js",
+    "code-style": "tsfmt --verify src/*.ts && jscs -p crockford test/*.js",
     "complexity": "node test/check-complexity.js",
     "static-analysis": "npm run check-version && npm run tslint && npm run code-style && npm run complexity",
     "unit-tests": "node test/unit-tests.js",
@@ -95,7 +96,7 @@
     "downstream": "node test/downstream.js",
     "travis": "npm test",
     "circleci": "npm test && npm run codecov && npm run downstream",
-    "appveyor": "npm run all-tests && npm run browser-tests && npm run dynamic-analysis",
+    "appveyor": "npm run compile && npm run all-tests && npm run browser-tests && npm run dynamic-analysis",
     "droneio": "npm test && npm run saucelabs-evergreen && npm run saucelabs-ie && npm run saucelabs-safari",
     "generate-regex": "node tools/generate-identifier-regex.js"
   }
