+++ src/scanner.ts
@@ -1,0 +1,1365 @@
+import { assert } from './assert';
+import { Messages} from './messages';
+import { Character} from './character';
+import { Token } from './token';
+
+function hexValue(ch: string): number {
+    return '0123456789abcdef'.indexOf(ch.toLowerCase());
+}
+
+function octalValue(ch: string): number {
+    return '01234567'.indexOf(ch);
+}
+
+export class Scanner {
+
+    source: string;
+    length: number;
+
+    strict: boolean;
+
+    index: number;
+    lineNumber: number;
+    lineStart: number;
+
+    startIndex: number;
+    startLineNumber: number;
+    startLineStart: number;
+
+    lastIndex: number;
+    lastLineNumber: number;
+    lastLineStart: number;
+
+    scanning: boolean;
+    lookahead: any;
+    hasLineTerminator: boolean;
+    curlyStack: string[];
+
+    stateStack: any[];
+
+    addComment = null;
+    addToken = null;
+    markBracket = null;
+    handleSlash = null;
+    throwUnexpectedToken: any;
+    tolerateUnexpectedToken: any;
+
+    constructor(code: any, delegate: any) {
+        if (typeof code !== 'string' && !(code instanceof String)) {
+            code = String(code);
+        }
+        this.source = code;
+        this.length = code.length;
+        this.strict = false;
+        this.index = 0;
+        this.lineNumber = (code.length > 0) ? 1 : 0;
+        this.lineStart = 0;
+        this.startIndex = this.lastIndex = 0;
+        this.startLineNumber = this.lastLineNumber = this.lineNumber;
+        this.startLineStart = this.lastLineStart = 0;
+        this.scanning = false;
+        this.lookahead = null;
+        this.hasLineTerminator = false;
+        this.curlyStack = [];
+        this.stateStack = [];
+
+        this.addComment = delegate.addComment;
+        this.throwUnexpectedToken = delegate.throwUnexpectedToken;
+        this.tolerateUnexpectedToken = delegate.tolerateUnexpectedToken;
+    };
+
+    pushState() {
+        this.stateStack.push({
+            strict: this.strict,
+            index: this.index,
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            startIndex: this.startIndex,
+            startLineNumber: this.startLineNumber,
+            startLineStart: this.startLineStart,
+            lastIndex: this.lastIndex,
+            lastLineNumber: this.lastLineNumber,
+            lastLineStart: this.lastLineStart,
+            scanning: this.scanning,
+            lookahead: this.lookahead,
+            hasLineTerminator: this.hasLineTerminator
+        });
+    };
+
+    popState() {
+        var state = this.stateStack.pop();
+        assert(typeof state === 'object', 'Scanner state is not balanced');
+
+        this.strict = state.strict;
+        this.index = state.index;
+        this.lineNumber = state.lineNumber;
+        this.lineStart = state.lineStart;
+        this.startIndex = state.startIndex;
+        this.startLineNumber = state.startLineNumber;
+        this.startLineStart = state.startLineStart;
+        this.lastIndex = state.lastIndex;
+        this.lastLineNumber = state.lastLineNumber;
+        this.lastLineStart = state.lastLineStart;
+        this.scanning = state.scanning;
+        this.lookahead = state.lookahead;
+        this.hasLineTerminator = state.hasLineTerminator;
+    }
+
+    eof(): boolean {
+        return this.index >= this.length;
+    };
+
+    // ECMA-262 11.4 Comments
+
+    skipSingleLineComment(offset: number): void {
+        var start, loc, ch, comment;
+
+        start = this.index - offset;
+        loc = {
+            start: {
+                line: this.lineNumber,
+                column: this.index - this.lineStart - offset
+            }
+        };
+
+        while (!this.eof()) {
+            ch = this.source.charCodeAt(this.index);
+            ++this.index;
+            if (Character.isLineTerminator(ch)) {
+                this.hasLineTerminator = true;
+                if (this.addComment) {
+                    comment = this.source.slice(start + offset, this.index - 1);
+                    loc.end = {
+                        line: this.lineNumber,
+                        column: this.index - this.lineStart - 1
+                    };
+                    this.addComment('Line', comment, start, this.index - 1, loc);
+                }
+                if (ch === 13 && this.source.charCodeAt(this.index) === 10) {
+                    ++this.index;
+                }
+                ++this.lineNumber;
+                this.lineStart = this.index;
+                return;
+            }
+        }
+
+        if (this.addComment) {
+            comment = this.source.slice(start + offset, this.index);
+            loc.end = {
+                line: this.lineNumber,
+                column: this.index - this.lineStart
+            };
+            this.addComment('Line', comment, start, this.index, loc);
+        }
+    };
+
+    skipMultiLineComment(): void {
+        var start, loc, ch, comment;
+
+        if (this.addComment) {
+            start = this.index - 2;
+            loc = {
+                start: {
+                    line: this.lineNumber,
+                    column: this.index - this.lineStart - 2
+                }
+            };
+        }
+
+        while (!this.eof()) {
+            ch = this.source.charCodeAt(this.index);
+            if (Character.isLineTerminator(ch)) {
+                if (ch === 0x0D && this.source.charCodeAt(this.index + 1) === 0x0A) {
+                    ++this.index;
+                }
+                this.hasLineTerminator = true;
+                ++this.lineNumber;
+                ++this.index;
+                this.lineStart = this.index;
+            } else if (ch === 0x2A) {
+                // Block comment ends with '*/'.
+                if (this.source.charCodeAt(this.index + 1) === 0x2F) {
+                    ++this.index;
+                    ++this.index;
+                    if (this.addComment) {
+                        comment = this.source.slice(start + 2, this.index - 2);
+                        loc.end = {
+                            line: this.lineNumber,
+                            column: this.index - this.lineStart
+                        };
+                        this.addComment('Block', comment, start, this.index, loc);
+                    }
+                    return;
+                }
+                ++this.index;
+            } else {
+                ++this.index;
+            }
+        }
+
+        // Ran off the end of the file - the whole thing is a comment
+        if (this.addComment) {
+            loc.end = {
+                line: this.lineNumber,
+                column: this.index - this.lineStart
+            };
+            comment = this.source.slice(start + 2, this.index);
+            this.addComment('Block', comment, start, this.index, loc);
+        }
+
+        this.tolerateUnexpectedToken();
+    };
+
+    skipComment(): void {
+        var ch, start;
+        this.hasLineTerminator = false;
+
+        start = (this.index === 0);
+        while (!this.eof()) {
+            ch = this.source.charCodeAt(this.index);
+
+            if (Character.isWhiteSpace(ch)) {
+                ++this.index;
+            } else if (Character.isLineTerminator(ch)) {
+                this.hasLineTerminator = true;
+                ++this.index;
+                if (ch === 0x0D && this.source.charCodeAt(this.index) === 0x0A) {
+                    ++this.index;
+                }
+                ++this.lineNumber;
+                this.lineStart = this.index;
+                start = true;
+            } else if (ch === 0x2F) { // U+002F is '/'
+                ch = this.source.charCodeAt(this.index + 1);
+                if (ch === 0x2F) {
+                    ++this.index;
+                    ++this.index;
+                    this.skipSingleLineComment(2);
+                    start = true;
+                } else if (ch === 0x2A) {  // U+002A is '*'
+                    ++this.index;
+                    ++this.index;
+                    this.skipMultiLineComment();
+                } else {
+                    break;
+                }
+            } else if (start && ch === 0x2D) { // U+002D is '-'
+                // U+003E is '>'
+                if ((this.source.charCodeAt(this.index + 1) === 0x2D) && (this.source.charCodeAt(this.index + 2) === 0x3E)) {
+                    // '-->' is a single-line comment
+                    this.index += 3;
+                    this.skipSingleLineComment(3);
+                } else {
+                    break;
+                }
+            } else if (ch === 0x3C) { // U+003C is '<'
+                if (this.source.slice(this.index + 1, this.index + 4) === '!--') {
+                    ++this.index; // `<`
+                    ++this.index; // `!`
+                    ++this.index; // `-`
+                    ++this.index; // `-`
+                    this.skipSingleLineComment(4);
+                } else {
+                    break;
+                }
+            } else {
+                break;
+            }
+        }
+    };
+
+    // ECMA-262 11.6.2.2 Future Reserved Words
+
+    isFutureReservedWord(id: string): boolean {
+        switch (id) {
+            case 'enum':
+            case 'export':
+            case 'import':
+            case 'super':
+                return true;
+            default:
+                return false;
+        }
+    };
+
+    isStrictModeReservedWord(id: string): boolean {
+        switch (id) {
+            case 'implements':
+            case 'interface':
+            case 'package':
+            case 'private':
+            case 'protected':
+            case 'public':
+            case 'static':
+            case 'yield':
+            case 'let':
+                return true;
+            default:
+                return false;
+        }
+    };
+
+    isRestrictedWord(id: string): boolean {
+        return id === 'eval' || id === 'arguments';
+    };
+
+    // ECMA-262 11.6.2.1 Keywords
+
+    isKeyword(id: string): boolean {
+        switch (id.length) {
+            case 2:
+                return (id === 'if') || (id === 'in') || (id === 'do');
+            case 3:
+                return (id === 'var') || (id === 'for') || (id === 'new') ||
+                    (id === 'try') || (id === 'let');
+            case 4:
+                return (id === 'this') || (id === 'else') || (id === 'case') ||
+                    (id === 'void') || (id === 'with') || (id === 'enum');
+            case 5:
+                return (id === 'while') || (id === 'break') || (id === 'catch') ||
+                    (id === 'throw') || (id === 'const') || (id === 'yield') ||
+                    (id === 'class') || (id === 'super');
+            case 6:
+                return (id === 'return') || (id === 'typeof') || (id === 'delete') ||
+                    (id === 'switch') || (id === 'export') || (id === 'import');
+            case 7:
+                return (id === 'default') || (id === 'finally') || (id === 'extends');
+            case 8:
+                return (id === 'function') || (id === 'continue') || (id === 'debugger');
+            case 10:
+                return (id === 'instanceof');
+            default:
+                return false;
+        }
+    };
+
+    codePointAt(i: number): number {
+        var cp, first, second;
+
+        cp = this.source.charCodeAt(i);
+        if (cp >= 0xD800 && cp <= 0xDBFF) {
+            second = this.source.charCodeAt(i + 1);
+            if (second >= 0xDC00 && second <= 0xDFFF) {
+                first = cp;
+                cp = (first - 0xD800) * 0x400 + second - 0xDC00 + 0x10000;
+            }
+        }
+
+        return cp;
+    };
+
+    scanHexEscape(prefix: string): string {
+        var i, len, code = 0;
+
+        len = (prefix === 'u') ? 4 : 2;
+        for (i = 0; i < len; ++i) {
+            if (!this.eof() && Character.isHexDigit(this.source.charCodeAt(this.index))) {
+                code = code * 16 + hexValue(this.source[this.index++]);
+            } else {
+                return '';
+            }
+        }
+        return String.fromCharCode(code);
+    };
+
+    scanUnicodeCodePointEscape(): string {
+        var ch, code;
+
+        ch = this.source[this.index];
+        code = 0;
+
+        // At least, one hex digit is required.
+        if (ch === '}') {
+            this.throwUnexpectedToken();
+        }
+
+        while (!this.eof()) {
+            ch = this.source[this.index++];
+            if (!Character.isHexDigit(ch.charCodeAt(0))) {
+                break;
+            }
+            code = code * 16 + hexValue(ch);
+        }
+
+        if (code > 0x10FFFF || ch !== '}') {
+            this.throwUnexpectedToken();
+        }
+
+        return Character.fromCodePoint(code);
+    };
+
+    getIdentifier(): string {
+        var start, ch;
+
+        start = this.index++;
+        while (!this.eof()) {
+            ch = this.source.charCodeAt(this.index);
+            if (ch === 0x5C) {
+                // Blackslash (U+005C) marks Unicode escape sequence.
+                this.index = start;
+                return this.getComplexIdentifier();
+            } else if (ch >= 0xD800 && ch < 0xDFFF) {
+                // Need to handle surrogate pairs.
+                this.index = start;
+                return this.getComplexIdentifier();
+            }
+            if (Character.isIdentifierPart(ch)) {
+                ++this.index;
+            } else {
+                break;
+            }
+        }
+
+        return this.source.slice(start, this.index);
+    };
+
+    getComplexIdentifier(): string {
+        var cp, ch, id;
+
+        cp = this.codePointAt(this.index);
+        id = Character.fromCodePoint(cp);
+        this.index += id.length;
+
+        // '\u' (U+005C, U+0075) denotes an escaped character.
+        if (cp === 0x5C) {
+            if (this.source.charCodeAt(this.index) !== 0x75) {
+                this.throwUnexpectedToken();
+            }
+            ++this.index;
+            if (this.source[this.index] === '{') {
+                ++this.index;
+                ch = this.scanUnicodeCodePointEscape();
+            } else {
+                ch = this.scanHexEscape('u');
+                cp = ch.charCodeAt(0);
+                if (!ch || ch === '\\' || !Character.isIdentifierStart(cp)) {
+                    this.throwUnexpectedToken();
+                }
+            }
+            id = ch;
+        }
+
+        while (!this.eof()) {
+            cp = this.codePointAt(this.index);
+            if (!Character.isIdentifierPart(cp)) {
+                break;
+            }
+            ch = Character.fromCodePoint(cp);
+            id += ch;
+            this.index += ch.length;
+
+            // '\u' (U+005C, U+0075) denotes an escaped character.
+            if (cp === 0x5C) {
+                id = id.substr(0, id.length - 1);
+                if (this.source.charCodeAt(this.index) !== 0x75) {
+                    this.throwUnexpectedToken();
+                }
+                ++this.index;
+                if (this.source[this.index] === '{') {
+                    ++this.index;
+                    ch = this.scanUnicodeCodePointEscape();
+                } else {
+                    ch = this.scanHexEscape('u');
+                    cp = ch.charCodeAt(0);
+                    if (!ch || ch === '\\' || !Character.isIdentifierPart(cp)) {
+                        this.throwUnexpectedToken();
+                    }
+                }
+                id += ch;
+            }
+        }
+
+        return id;
+    };
+
+    octalToDecimal(ch: string) {
+        // \0 is not octal escape sequence
+        var octal = (ch !== '0'), code = octalValue(ch);
+
+        if (!this.eof() && Character.isOctalDigit(this.source.charCodeAt(this.index))) {
+            octal = true;
+            code = code * 8 + octalValue(this.source[this.index++]);
+
+            // 3 digits are only allowed when string starts
+            // with 0, 1, 2, 3
+            if ('0123'.indexOf(ch) >= 0 && !this.eof() && Character.isOctalDigit(this.source.charCodeAt(this.index))) {
+                code = code * 8 + octalValue(this.source[this.index++]);
+            }
+        }
+
+        return {
+            code: code,
+            octal: octal
+        };
+    };
+
+    // ECMA-262 11.6 Names and Keywords
+
+    scanIdentifier() {
+        var start, id, type;
+
+        start = this.index;
+
+        // Backslash (U+005C) starts an escaped character.
+        start = this.index;
+        id = (this.source.charCodeAt(this.index) === 0x5C) ? this.getComplexIdentifier() : this.getIdentifier();
+
+        // There is no keyword or literal with only one character.
+        // Thus, it must be an identifier.
+        if (id.length === 1) {
+            type = Token.Identifier;
+        } else if (this.isKeyword(id)) {
+            type = Token.Keyword;
+        } else if (id === 'null') {
+            type = Token.NullLiteral;
+        } else if (id === 'true' || id === 'false') {
+            type = Token.BooleanLiteral;
+        } else {
+            type = Token.Identifier;
+        }
+
+        return {
+            type: type,
+            value: id,
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    // ECMA-262 11.7 Punctuators
+
+    scanPunctuator() {
+        var token, str;
+
+        token = {
+            type: Token.Punctuator,
+            value: '',
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: this.index,
+            end: this.index
+        };
+
+        // Check for most common single-character punctuators.
+        str = this.source[this.index];
+        switch (str) {
+
+            case '(':
+            case '{':
+                if (this.markBracket) {
+                    this.markBracket(str);
+                }
+                if (str === '{') {
+                    this.curlyStack.push('{');
+                }
+                ++this.index;
+                break;
+
+            case '.':
+                ++this.index;
+                if (this.source[this.index] === '.' && this.source[this.index + 1] === '.') {
+                    // Spread operator: ...
+                    this.index += 2;
+                    str = '...';
+                }
+                break;
+
+            case '}':
+                ++this.index;
+                this.curlyStack.pop();
+                break;
+            case ')':
+            case ';':
+            case ',':
+            case '[':
+            case ']':
+            case ':':
+            case '?':
+            case '~':
+                ++this.index;
+                break;
+
+            default:
+                // 4-character punctuator.
+                str = this.source.substr(this.index, 4);
+                if (str === '>>>=') {
+                    this.index += 4;
+                } else {
+
+                    // 3-character punctuators.
+                    str = str.substr(0, 3);
+                    if (str === '===' || str === '!==' || str === '>>>' ||
+                        str === '<<=' || str === '>>=') {
+                        this.index += 3;
+                    } else {
+
+                        // 2-character punctuators.
+                        str = str.substr(0, 2);
+                        if (str === '&&' || str === '||' || str === '==' || str === '!=' ||
+                            str === '+=' || str === '-=' || str === '*=' || str === '/=' ||
+                            str === '++' || str === '--' || str === '<<' || str === '>>' ||
+                            str === '&=' || str === '|=' || str === '^=' || str === '%=' ||
+                            str === '<=' || str === '>=' || str === '=>') {
+                            this.index += 2;
+                        } else {
+
+                            // 1-character punctuators.
+                            str = this.source[this.index];
+                            if ('<>=!+-*%&|^/'.indexOf(str) >= 0) {
+                                ++this.index;
+                            }
+                        }
+                    }
+                }
+        }
+
+        if (this.index === token.start) {
+            this.throwUnexpectedToken();
+        }
+
+        token.end = this.index;
+        token.value = str;
+        return token;
+    };
+
+    // ECMA-262 11.8.3 Numeric Literals
+
+    scanHexLiteral(start: number) {
+        var number = '';
+
+        while (!this.eof()) {
+            if (!Character.isHexDigit(this.source.charCodeAt(this.index))) {
+                break;
+            }
+            number += this.source[this.index++];
+        }
+
+        if (number.length === 0) {
+            this.throwUnexpectedToken();
+        }
+
+        if (Character.isIdentifierStart(this.source.charCodeAt(this.index))) {
+            this.throwUnexpectedToken();
+        }
+
+        return {
+            type: Token.NumericLiteral,
+            value: parseInt('0x' + number, 16),
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    scanBinaryLiteral(start: number) {
+        var ch, number;
+
+        number = '';
+
+        while (!this.eof()) {
+            ch = this.source[this.index];
+            if (ch !== '0' && ch !== '1') {
+                break;
+            }
+            number += this.source[this.index++];
+        }
+
+        if (number.length === 0) {
+            // only 0b or 0B
+            this.throwUnexpectedToken();
+        }
+
+        if (!this.eof()) {
+            ch = this.source.charCodeAt(this.index);
+            /* istanbul ignore else */
+            if (Character.isIdentifierStart(ch) || Character.isDecimalDigit(ch)) {
+                this.throwUnexpectedToken();
+            }
+        }
+
+        return {
+            type: Token.NumericLiteral,
+            value: parseInt(number, 2),
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    scanOctalLiteral(prefix: string, start: number) {
+        var number, octal;
+
+        if (Character.isOctalDigit(prefix.charCodeAt(0))) {
+            octal = true;
+            number = '0' + this.source[this.index++];
+        } else {
+            octal = false;
+            ++this.index;
+            number = '';
+        }
+
+        while (!this.eof()) {
+            if (!Character.isOctalDigit(this.source.charCodeAt(this.index))) {
+                break;
+            }
+            number += this.source[this.index++];
+        }
+
+        if (!octal && number.length === 0) {
+            // only 0o or 0O
+            this.throwUnexpectedToken();
+        }
+
+        if (Character.isIdentifierStart(this.source.charCodeAt(this.index)) || Character.isDecimalDigit(this.source.charCodeAt(this.index))) {
+            this.throwUnexpectedToken();
+        }
+
+        return {
+            type: Token.NumericLiteral,
+            value: parseInt(number, 8),
+            octal: octal,
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    isImplicitOctalLiteral(): boolean {
+        var i, ch;
+
+        // Implicit octal, unless there is a non-octal digit.
+        // (Annex B.1.1 on Numeric Literals)
+        for (i = this.index + 1; i < this.length; ++i) {
+            ch = this.source[i];
+            if (ch === '8' || ch === '9') {
+                return false;
+            }
+            if (!Character.isOctalDigit(ch.charCodeAt(0))) {
+                return true;
+            }
+        }
+
+        return true;
+    };
+
+    scanNumericLiteral() {
+        var number, start, ch;
+
+        ch = this.source[this.index];
+        assert(Character.isDecimalDigit(ch.charCodeAt(0)) || (ch === '.'),
+            'Numeric literal must start with a decimal digit or a decimal point');
+
+        start = this.index;
+        number = '';
+        if (ch !== '.') {
+            number = this.source[this.index++];
+            ch = this.source[this.index];
+
+            // Hex number starts with '0x'.
+            // Octal number starts with '0'.
+            // Octal number in ES6 starts with '0o'.
+            // Binary number in ES6 starts with '0b'.
+            if (number === '0') {
+                if (ch === 'x' || ch === 'X') {
+                    ++this.index;
+                    return this.scanHexLiteral(start);
+                }
+                if (ch === 'b' || ch === 'B') {
+                    ++this.index;
+                    return this.scanBinaryLiteral(start);
+                }
+                if (ch === 'o' || ch === 'O') {
+                    return this.scanOctalLiteral(ch, start);
+                }
+
+                if (ch && Character.isOctalDigit(ch.charCodeAt(0))) {
+                    if (this.isImplicitOctalLiteral()) {
+                        return this.scanOctalLiteral(ch, start);
+                    }
+                }
+            }
+
+            while (Character.isDecimalDigit(this.source.charCodeAt(this.index))) {
+                number += this.source[this.index++];
+            }
+            ch = this.source[this.index];
+        }
+
+        if (ch === '.') {
+            number += this.source[this.index++];
+            while (Character.isDecimalDigit(this.source.charCodeAt(this.index))) {
+                number += this.source[this.index++];
+            }
+            ch = this.source[this.index];
+        }
+
+        if (ch === 'e' || ch === 'E') {
+            number += this.source[this.index++];
+
+            ch = this.source[this.index];
+            if (ch === '+' || ch === '-') {
+                number += this.source[this.index++];
+            }
+            if (Character.isDecimalDigit(this.source.charCodeAt(this.index))) {
+                while (Character.isDecimalDigit(this.source.charCodeAt(this.index))) {
+                    number += this.source[this.index++];
+                }
+            } else {
+                this.throwUnexpectedToken();
+            }
+        }
+
+        if (Character.isIdentifierStart(this.source.charCodeAt(this.index))) {
+            this.throwUnexpectedToken();
+        }
+
+        return {
+            type: Token.NumericLiteral,
+            value: parseFloat(number),
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    // ECMA-262 11.8.4 String Literals
+
+    scanStringLiteral() {
+        var str = '', quote, start, ch, unescaped, octToDec, octal = false;
+
+        quote = this.source[this.index];
+        assert((quote === '\'' || quote === '"'),
+            'String literal must starts with a quote');
+
+        start = this.index;
+        ++this.index;
+
+        while (!this.eof()) {
+            ch = this.source[this.index++];
+
+            if (ch === quote) {
+                quote = '';
+                break;
+            } else if (ch === '\\') {
+                ch = this.source[this.index++];
+                if (!ch || !Character.isLineTerminator(ch.charCodeAt(0))) {
+                    switch (ch) {
+                        case 'u':
+                        case 'x':
+                            if (this.source[this.index] === '{') {
+                                ++this.index;
+                                str += this.scanUnicodeCodePointEscape();
+                            } else {
+                                unescaped = this.scanHexEscape(ch);
+                                if (!unescaped) {
+                                    this.throwUnexpectedToken();
+                                }
+                                str += unescaped;
+                            }
+                            break;
+                        case 'n':
+                            str += '\n';
+                            break;
+                        case 'r':
+                            str += '\r';
+                            break;
+                        case 't':
+                            str += '\t';
+                            break;
+                        case 'b':
+                            str += '\b';
+                            break;
+                        case 'f':
+                            str += '\f';
+                            break;
+                        case 'v':
+                            str += '\x0B';
+                            break;
+                        case '8':
+                        case '9':
+                            str += ch;
+                            this.tolerateUnexpectedToken();
+                            break;
+
+                        default:
+                            if (ch && Character.isOctalDigit(ch.charCodeAt(0))) {
+                                octToDec = this.octalToDecimal(ch);
+
+                                octal = octToDec.octal || octal;
+                                str += String.fromCharCode(octToDec.code);
+                            } else {
+                                str += ch;
+                            }
+                            break;
+                    }
+                } else {
+                    ++this.lineNumber;
+                    if (ch === '\r' && this.source[this.index] === '\n') {
+                        ++this.index;
+                    }
+                    this.lineStart = this.index;
+                }
+            } else if (Character.isLineTerminator(ch.charCodeAt(0))) {
+                break;
+            } else {
+                str += ch;
+            }
+        }
+
+        if (quote !== '') {
+            this.throwUnexpectedToken();
+        }
+
+        return {
+            type: Token.StringLiteral,
+            value: str,
+            octal: octal,
+            lineNumber: this.startLineNumber,
+            lineStart: this.startLineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    // ECMA-262 11.8.6 Template Literal Lexical Components
+
+    scanTemplate() {
+        var cooked = '', ch, start, rawOffset, terminated, head, tail, restore, unescaped;
+
+        terminated = false;
+        tail = false;
+        start = this.index;
+        head = (this.source[this.index] === '`');
+        rawOffset = 2;
+
+        ++this.index;
+
+        while (!this.eof()) {
+            ch = this.source[this.index++];
+            if (ch === '`') {
+                rawOffset = 1;
+                tail = true;
+                terminated = true;
+                break;
+            } else if (ch === '$') {
+                if (this.source[this.index] === '{') {
+                    this.curlyStack.push('${');
+                    ++this.index;
+                    terminated = true;
+                    break;
+                }
+                cooked += ch;
+            } else if (ch === '\\') {
+                ch = this.source[this.index++];
+                if (!Character.isLineTerminator(ch.charCodeAt(0))) {
+                    switch (ch) {
+                        case 'n':
+                            cooked += '\n';
+                            break;
+                        case 'r':
+                            cooked += '\r';
+                            break;
+                        case 't':
+                            cooked += '\t';
+                            break;
+                        case 'u':
+                        case 'x':
+                            if (this.source[this.index] === '{') {
+                                ++this.index;
+                                cooked += this.scanUnicodeCodePointEscape();
+                            } else {
+                                restore = this.index;
+                                unescaped = this.scanHexEscape(ch);
+                                if (unescaped) {
+                                    cooked += unescaped;
+                                } else {
+                                    this.index = restore;
+                                    cooked += ch;
+                                }
+                            }
+                            break;
+                        case 'b':
+                            cooked += '\b';
+                            break;
+                        case 'f':
+                            cooked += '\f';
+                            break;
+                        case 'v':
+                            cooked += '\v';
+                            break;
+
+                        default:
+                            if (ch === '0') {
+                                if (Character.isDecimalDigit(this.source.charCodeAt(this.index))) {
+                                    // Illegal: \01 \02 and so on
+                                    this.throwUnexpectedToken(null, Messages.TemplateOctalLiteral);
+                                }
+                                cooked += '\0';
+                            } else if (Character.isOctalDigit(ch.charCodeAt(0))) {
+                                // Illegal: \1 \2
+                                this.throwUnexpectedToken(null, Messages.TemplateOctalLiteral);
+                            } else {
+                                cooked += ch;
+                            }
+                            break;
+                    }
+                } else {
+                    ++this.lineNumber;
+                    if (ch === '\r' && this.source[this.index] === '\n') {
+                        ++this.index;
+                    }
+                    this.lineStart = this.index;
+                }
+            } else if (Character.isLineTerminator(ch.charCodeAt(0))) {
+                ++this.lineNumber;
+                if (ch === '\r' && this.source[this.index] === '\n') {
+                    ++this.index;
+                }
+                this.lineStart = this.index;
+                cooked += '\n';
+            } else {
+                cooked += ch;
+            }
+        }
+
+        if (!terminated) {
+            this.throwUnexpectedToken();
+        }
+
+        if (!head) {
+            this.curlyStack.pop();
+        }
+
+        return {
+            type: Token.Template,
+            value: {
+                cooked: cooked,
+                raw: this.source.slice(start + 1, this.index - rawOffset)
+            },
+            head: head,
+            tail: tail,
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    // ECMA-262 11.8.5 Regular Expression Literals
+
+    testRegExp(pattern: string, flags: string) {
+        // The BMP character to use as a replacement for astral symbols when
+        // translating an ES6 "u"-flagged pattern to an ES5-compatible
+        // approximation.
+        // Note: replacing with '\uFFFF' enables false positives in unlikely
+        // scenarios. For example, `[\u{1044f}-\u{10440}]` is an invalid
+        // pattern that would not be detected by this substitution.
+        var astralSubstitute = '\uFFFF',
+            tmp = pattern,
+            self = this;
+
+        if (flags.indexOf('u') >= 0) {
+            tmp = tmp
+            // Replace every Unicode escape sequence with the equivalent
+            // BMP character or a constant ASCII code point in the case of
+            // astral symbols. (See the above note on `astralSubstitute`
+            // for more information.)
+                .replace(/\\u\{([0-9a-fA-F]+)\}|\\u([a-fA-F0-9]{4})/g, function($0, $1, $2) {
+                    var codePoint = parseInt($1 || $2, 16);
+                    if (codePoint > 0x10FFFF) {
+                        self.throwUnexpectedToken(null, Messages.InvalidRegExp);
+                    }
+                    if (codePoint <= 0xFFFF) {
+                        return String.fromCharCode(codePoint);
+                    }
+                    return astralSubstitute;
+                })
+            // Replace each paired surrogate with a single ASCII symbol to
+            // avoid throwing on regular expressions that are only valid in
+            // combination with the "u" flag.
+                .replace(
+                /[\uD800-\uDBFF][\uDC00-\uDFFF]/g,
+                astralSubstitute
+                );
+        }
+
+        // First, detect invalid regular expressions.
+        try {
+            RegExp(tmp);
+        } catch (e) {
+            this.throwUnexpectedToken(null, Messages.InvalidRegExp);
+        }
+
+        // Return a regular expression object for this pattern-flag pair, or
+        // `null` in case the current environment doesn't support the flags it
+        // uses.
+        try {
+            return new RegExp(pattern, flags);
+        } catch (exception) {
+            return null;
+        }
+    };
+
+    scanRegExpBody() {
+        var ch, str, classMarker, terminated, body;
+
+        ch = this.source[this.index];
+        assert(ch === '/', 'Regular expression literal must start with a slash');
+        str = this.source[this.index++];
+
+        classMarker = false;
+        terminated = false;
+        while (!this.eof()) {
+            ch = this.source[this.index++];
+            str += ch;
+            if (ch === '\\') {
+                ch = this.source[this.index++];
+                // ECMA-262 7.8.5
+                if (Character.isLineTerminator(ch.charCodeAt(0))) {
+                    this.throwUnexpectedToken(null, Messages.UnterminatedRegExp);
+                }
+                str += ch;
+            } else if (Character.isLineTerminator(ch.charCodeAt(0))) {
+                this.throwUnexpectedToken(null, Messages.UnterminatedRegExp);
+            } else if (classMarker) {
+                if (ch === ']') {
+                    classMarker = false;
+                }
+            } else {
+                if (ch === '/') {
+                    terminated = true;
+                    break;
+                } else if (ch === '[') {
+                    classMarker = true;
+                }
+            }
+        }
+
+        if (!terminated) {
+            this.throwUnexpectedToken(null, Messages.UnterminatedRegExp);
+        }
+
+        // Exclude leading and trailing slash.
+        body = str.substr(1, str.length - 2);
+        return {
+            value: body,
+            literal: str
+        };
+    };
+
+    scanRegExpFlags() {
+        var ch, str, flags, restore;
+
+        str = '';
+        flags = '';
+        while (!this.eof()) {
+            ch = this.source[this.index];
+            if (!Character.isIdentifierPart(ch.charCodeAt(0))) {
+                break;
+            }
+
+            ++this.index;
+            if (ch === '\\' && !this.eof()) {
+                ch = this.source[this.index];
+                if (ch === 'u') {
+                    ++this.index;
+                    restore = this.index;
+                    ch = this.scanHexEscape('u');
+                    if (ch) {
+                        flags += ch;
+                        for (str += '\\u'; restore < this.index; ++restore) {
+                            str += this.source[restore];
+                        }
+                    } else {
+                        this.index = restore;
+                        flags += 'u';
+                        str += '\\u';
+                    }
+                    this.tolerateUnexpectedToken();
+                } else {
+                    str += '\\';
+                    this.tolerateUnexpectedToken();
+                }
+            } else {
+                flags += ch;
+                str += ch;
+            }
+        }
+
+        return {
+            value: flags,
+            literal: str
+        };
+    };
+
+    parseRegExpLiteral() {
+        var start, body, flags, value;
+
+        this.scanning = true;
+        this.lookahead = null;
+        this.skipComment();
+        start = this.index;
+
+        body = this.scanRegExpBody();
+        flags = this.scanRegExpFlags();
+        value = this.testRegExp(body.value, flags.value);
+        this.scanning = false;
+
+        return {
+            literal: body.literal + flags.literal,
+            value: value,
+            regex: {
+                pattern: body.value,
+                flags: flags.value
+            },
+            start: start,
+            end: this.index
+        };
+    };
+
+    scanRegExp() {
+        var start, body, flags, value;
+
+        this.scanning = true;
+        this.lookahead = null;
+        this.skipComment();
+        start = this.index;
+
+        body = this.scanRegExpBody();
+        flags = this.scanRegExpFlags();
+        value = this.testRegExp(body.value, flags.value);
+        this.scanning = false;
+
+        return {
+            type: Token.RegularExpression,
+            value: value,
+            regex: {
+                pattern: body.value,
+                flags: flags.value
+            },
+            lineNumber: this.lineNumber,
+            lineStart: this.lineStart,
+            start: start,
+            end: this.index
+        };
+    };
+
+    advance() {
+        var cp, token;
+
+        if (this.eof()) {
+            return {
+                type: Token.EOF,
+                lineNumber: this.lineNumber,
+                lineStart: this.lineStart,
+                start: this.index,
+                end: this.index
+            };
+        }
+
+        cp = this.source.charCodeAt(this.index);
+
+        if (Character.isIdentifierStart(cp)) {
+            token = this.scanIdentifier();
+            if (this.strict && this.isStrictModeReservedWord(token.value)) {
+                token.type = Token.Keyword;
+            }
+            return token;
+        }
+
+        // Very common: ( and ) and ;
+        if (cp === 0x28 || cp === 0x29 || cp === 0x3B) {
+            return this.scanPunctuator();
+        }
+
+        // String literal starts with single quote (U+0027) or double quote (U+0022).
+        if (cp === 0x27 || cp === 0x22) {
+            return this.scanStringLiteral();
+        }
+
+        // Dot (.) U+002E can also start a floating-point number, hence the need
+        // to check the next character.
+        if (cp === 0x2E) {
+            if (Character.isDecimalDigit(this.source.charCodeAt(this.index + 1))) {
+                return this.scanNumericLiteral();
+            }
+            return this.scanPunctuator();
+        }
+
+        if (Character.isDecimalDigit(cp)) {
+            return this.scanNumericLiteral();
+        }
+
+        // Slash (/) U+002F can also start a regex.
+        if (cp === 0x2F && this.handleSlash) {
+            return this.handleSlash();
+        }
+
+        // Template literals start with ` (U+0060) for template head
+        // or } (U+007D) for template middle or template tail.
+        if (cp === 0x60 || (cp === 0x7D && this.curlyStack[this.curlyStack.length - 1] === '${')) {
+            return this.scanTemplate();
+        }
+
+        // Possible identifier start in a surrogate pair.
+        if (cp >= 0xD800 && cp < 0xDFFF) {
+            cp = this.codePointAt(this.index);
+            if (Character.isIdentifierStart(cp)) {
+                return this.scanIdentifier();
+            }
+        }
+
+        return this.scanPunctuator();
+    };
+
+    start() {
+        this.scanning = true;
+
+        this.skipComment();
+
+        this.startIndex = this.lastIndex = this.index;
+        this.startLineNumber = this.lastLineNumber = this.lineNumber;
+        this.startLineStart = this.lastLineStart = this.lineStart;
+
+        this.lookahead = this.advance();
+        if (this.addToken) {
+            this.addToken(this.lookahead);
+        }
+
+        this.scanning = false;
+    };
+
+
+    lex() {
+        var token;
+        this.scanning = true;
+
+        this.lastIndex = this.index;
+        this.lastLineNumber = this.lineNumber;
+        this.lastLineStart = this.lineStart;
+
+        this.skipComment();
+
+        token = this.lookahead;
+
+        this.startIndex = this.index;
+        this.startLineNumber = this.lineNumber;
+        this.startLineStart = this.lineStart;
+
+        this.lookahead = this.advance();
+        if (this.addToken) {
+            this.addToken(this.lookahead);
+        }
+
+        this.scanning = false;
+        return token;
+    };
+
+}
