+++ src/token.ts
@@ -1,0 +1,12 @@
+export enum Token {
+    BooleanLiteral = 1,
+    EOF,
+    Identifier,
+    Keyword,
+    NullLiteral,
+    NumericLiteral,
+    Punctuator,
+    StringLiteral,
+    RegularExpression,
+    Template
+};
